LayoutTests/http/tests/websocket/tests/hybi/bufferedAmount-after-close-expected.txt

Web Socket bufferedAmount after closed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
Closed.
PASS ws.readyState is 3
PASS ws.bufferedAmount is 0
Testing send(string)...
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 27
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 6
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 7
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 131
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 134
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 65543
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 65550
Testing send(ArrayBuffer)...
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 6
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 7
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 131
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 134
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 65543
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 65550
Testing send(Blob)...
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 6
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 7
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 131
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 134
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 65543
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 65550
PASS successfullyParsed is true

TEST COMPLETE
